feat: :art: createAsset refactoring
