+++ src/components/Wizard/createAsset/createNFT/index.tsx
@@ -5,27 +5,24 @@ import { parseAddress } from '@/utils/parseValues';
 import { useTranslation } from 'next-i18next';
 import { PropsWithChildren, useEffect, useState } from 'react';
 import { FormProvider, useForm } from 'react-hook-form';
-import {
-  ConfirmSuccessTransaction,
-  ConfirmTransaction,
-  CreateAssetAddRoles,
-  CreateAssetEightStep,
-  CreateAssetMaxSupplyStep,
-  CreateAssetNameStep,
-  CreateAssetOwnerAddressStep,
-  CreateAssetPreConfimStep,
-  CreateAssetPropertiesStep,
-  CreateAssetRoyaltySteps,
-  CreateAssetTickerStep,
-  CreateAssetWelcomeStep,
-  DesktopStepsComponent,
-  StepsBasics,
-  URIsSection,
-  infinitySymbol,
-  propertiesCommonDefaultValues,
-} from '..';
+import { infinitySymbol, propertiesCommonDefaultValues } from '..';
 import { createNFT } from '../../utils';
+import { ConfirmSuccessTransaction } from '../ConfirmSuccessTransaction';
+import { ConfirmTransaction } from '../ConfirmTransaction';
+import { CreateAssetAddRoles } from '../CreateAssetAddRoles';
+import { CreateAssetEightStep } from '../CreateAssetEightStep';
+import { CreateAssetMaxSupplyStep } from '../CreateAssetMaxSupplyStep';
+import { CreateAssetNameStep } from '../CreateAssetNameStep';
+import { CreateAssetOwnerAddressStep } from '../CreateAssetOwnerAddressStep';
+import { CreateAssetPreConfimStep } from '../CreateAssetPreConfimStep';
+import { CreateAssetPropertiesStep } from '../CreateAssetPropertiesStep';
+import { CreateAssetRoyaltySteps } from '../CreateAssetRoyaltySteps';
+import { CreateAssetTickerStep } from '../CreateAssetTickerStep';
+import { CreateAssetWelcomeStep } from '../CreateAssetWelcomeStep';
+import { DesktopStepsComponent } from '../DesktopStepsComponent';
+import { StepsBasics } from '../StepsBasics';
 import { WizardBody } from '../styles';
+import { URIsSection } from '../URIsSection';
 
 const WizCreateNFT: React.FC<PropsWithChildren<any>> = ({
   setTxHash,
